+++ PyDodo/pydodo/__init__.py
@@ -1,16 +1,16 @@
 
 from .request_position import aircraft_position, all_positions
-from .request_change import (
+from .aircraft_control import (
     change_altitude,
     change_heading,
     change_speed,
-    change_vertical_speed,
+    change_vertical_speed
 )
 from .create_aircraft import create_aircraft
 from .load_scenario import load_scenario
-from .manage_simulation import (
+from .simulation_control import (
     reset_simulation,
     pause_simulation,
     resume_simulation,
-    set_simulation_rate_multiplier,
+    set_simulation_rate_multiplier
 )
